requirements.txt
streamlit
google-generativeai
python-docx
python-pptx
Pillow
watchdog
